requirements.txt
python-dotenv==1.0.0
gooddata_sdk==1.4.0
